[Static Runtime] Cleanup LLVMCodeGen memory after code gen completes (#66218)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66218

This stack of diffs reduces the memory used by LLVMCodeGen object.

Here are the numbers on model `294738512`: (this is the number reported as `Memory turnover after freeze_module:` in the output)

```
Before: 123343496
After : 121566008
```

So, there is a reduction of about `~1.77MB` with this change of making `PytorchLLVMJIT` a singleton.

Test Plan: Imported from OSS

Reviewed By: ZolotukhinM, hlu1

Differential Revision: D31445798

Pulled By: navahgar

fbshipit-source-id: c860d36456b2c5d3e21010c1217e2948326f666d