Add support for cat in output stitching (#66098)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66098

`cat` is somewhat special-cased right now because currently we only have list of Tensor inputs where the list is constructed in the JIT IR graph. While that is generally true for Fusion (e.g. why we have ConstantChunk) that may not be true for shape analysis generally, so I'm waiting a bit to generalize.

Test Plan: Imported from OSS

Reviewed By: navahgar

Differential Revision: D31732415

Pulled By: eellison

fbshipit-source-id: 7f513cea355f1e4c1d2ca7c32c06690a9bdcb050